Publishing tool 50% functionality
Got it in my head that committing early with little progress would look worse than committing nothing until the last minute. Lots of foundational logic in place, as well as a UI that provides dynamic options to user and detects the specs of the open file. Some detection of user error to prevent issues in adhering to pipeline. Yet to read/write files, but formatting and browsing is there.